requirements.txt
requests==2.28.2
lxml==4.9.2
python-telegram-bot==13.7
python-dotenv==1.0.0